requirements.txt
amqp==5.0.9
argcomplete==1.12.3
attrs==21.4.0
billiard==3.6.4.0
bullet==2.2.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coloredlogs==15.0.1
docker==5.0.3
grpcio==1.44.0
grpcio-tools==1.44.0
humanfriendly==10.0
idna==3.3
importlib-metadata==4.11.1
iniconfig==1.1.1
kombu==5.2.3
lark==1.1.1
miniwdl==1.5.0
packaging==21.3
pluggy==1.0.0
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
py==1.11.0
pygtail==0.11.1
pyparsing==3.0.7
pytest==7.0.1
python-json-logger==2.0.2
pytz==2021.3
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.2.3
xdg==5.1.1
xtermcolor==1.3
zipp==3.7.0
